Crate crseo

CEO wrapper crate

The CEO wrapper is the interface to CEO CUDA API. CEO elements are created using the builder associated to each element.

For example, the default CEO elements Gmt and Source are built with:

use ceo::ceo;
let mut gmt = ceo!(GMT);
let mut src = ceo!(SOURCE);
src.through(&mut gmt).xpupil();
println!("WFE RMS: {:?}nm",src.wfe_rms_10e(-9));

ceo! is a macro that incorporates the necessary boilerplate code to create CEO elements.

Modules

analytic
Analytic Ray Tracing

atmosphere

calibrations

centroiding

ceo_bindings

cu

dos
DOS Interface

error

fwhm

gmt
CEO gmt wrapper

imaging

lmmse

pssn

raytracing

shackhartmann
CEO shackhartmann wrapper

source
CEO source wrapper

Macros

ceo
CEO macro builder

Structs

ATMOSPHERE
CEO Atmosphere builder type

Atmosphere

Calibration
GMT segment rigid body motion and surface figure calibration

Centroiding
Wrapper for CEO centroiding

CeoError

Cu

Fwhm

GMT
Gmt builder

Gmt
gmt wrapper

Imaging
An optical imager with a detector

LMMSE

LinearMinimumMeanSquareError

Mask

PSSN
CEO PSSn builder type

PSSn

SH48
ShackHartmann “SH48” builder for GMT AGWS model

SHACKHARTMANN
ShackHartmann builder

SOURCE
Source builder

ShackHartmann
shackhartmann wrapper

Source
source wrapper

Enums

CrseoError

Traits

Builder
CEO builder type trait

Propagation
A system that mutates Source arguments should implement the Propagation trait

Functions

set_gpu

Type Definitions

Diffractive

Geometric

GeometricShackHartmann

Result