Module crseo::ceo_bindings

Re-exports

pub use self::cudaError as cudaError_t;

Structs

BTBT

CUevent_st

GBTBT

LMMSE

aaStats

aperture

atmosphere

bundle

centroiding

complex_amplitude

conic

coordinate_system

cublasContext

curandStateXORWOW
CURAND XORWOW state

cusolverDnContext

cusparseContext

cusparseMatDescr

float2

geometricShackHartmann

gmt_m1

gmt_m2

gpu_double

gpu_float

imaging

iterativeSolvers

layer

mask

modes

paStats

profile

pssn

pyramid

ray

shackHartmann

source

stats

stopwatch

vector

zernikeS

Constants

cublasStatus_t_CUBLAS_STATUS_ALLOC_FAILED

cublasStatus_t_CUBLAS_STATUS_ARCH_MISMATCH

cublasStatus_t_CUBLAS_STATUS_EXECUTION_FAILED

cublasStatus_t_CUBLAS_STATUS_INTERNAL_ERROR

cublasStatus_t_CUBLAS_STATUS_INVALID_VALUE

cublasStatus_t_CUBLAS_STATUS_LICENSE_ERROR

cublasStatus_t_CUBLAS_STATUS_MAPPING_ERROR

cublasStatus_t_CUBLAS_STATUS_NOT_INITIALIZED

cublasStatus_t_CUBLAS_STATUS_NOT_SUPPORTED

cublasStatus_t_CUBLAS_STATUS_SUCCESS

cudaError_cudaErrorAddressOfConstant
This indicated that the user has taken the address of a constant variable, which was forbidden up until the CUDA 3.1 release. \deprecated This error return is deprecated as of CUDA 3.1. Variables in constant memory may now have their address taken by the runtime via ::cudaGetSymbolAddress().

cudaError_cudaErrorAlreadyAcquired
This indicates that a resource has already been acquired.

cudaError_cudaErrorAlreadyMapped
This indicates that the resource is already mapped.

cudaError_cudaErrorApiFailureBase
Any unhandled CUDA driver error is added to this value and returned via the runtime. Production releases of CUDA should not return such errors. \deprecated This error return is deprecated as of CUDA 4.1.

cudaError_cudaErrorArrayIsMapped
This indicates that the specified array is currently mapped and thus cannot be destroyed.

cudaError_cudaErrorAssert
An assert triggered in device code during kernel execution. The device cannot be used again. All existing allocations are invalid. To continue using CUDA, the process must be terminated and relaunched.

cudaError_cudaErrorCapturedEvent
The operation is not permitted on an event which was last recorded in a capturing stream.

cudaError_cudaErrorCompatNotSupportedOnDevice
This error indicates that the system was upgraded to run with forward compatibility but the visible hardware detected by CUDA does not support this configuration. Refer to the compatibility documentation for the supported hardware matrix or ensure that only supported hardware is visible during initialization via the CUDA_VISIBLE_DEVICES environment variable.

cudaError_cudaErrorContextIsDestroyed
This error indicates that the context current to the calling thread has been destroyed using ::cuCtxDestroy, or is a primary context which has not yet been initialized.

cudaError_cudaErrorCooperativeLaunchTooLarge
This error indicates that the number of blocks launched per grid for a kernel that was launched via either ::cudaLaunchCooperativeKernel or ::cudaLaunchCooperativeKernelMultiDevice exceeds the maximum number of blocks as allowed by ::cudaOccupancyMaxActiveBlocksPerMultiprocessor or ::cudaOccupancyMaxActiveBlocksPerMultiprocessorWithFlags times the number of multiprocessors as specified by the device attribute ::cudaDevAttrMultiProcessorCount.

cudaError_cudaErrorCudartUnloading
This indicates that a CUDA Runtime API call cannot be executed because it is being called during process shut down, at a point in time after CUDA driver has been unloaded.

cudaError_cudaErrorDeviceAlreadyInUse
This indicates that a call tried to access an exclusive-thread device that is already in use by a different thread.

cudaError_cudaErrorDeviceUninitilialized
This most frequently indicates that there is no context bound to the current thread. This can also be returned if the context passed to an API call is not a valid handle (such as a context that has had ::cuCtxDestroy() invoked on it). This can also be returned if a user mixes different API versions (i.e. 3010 context with 3020 API calls). See ::cuCtxGetApiVersion() for more details.

cudaError_cudaErrorDevicesUnavailable
This indicates that all CUDA devices are busy or unavailable at the current time. Devices are often busy/unavailable due to use of ::cudaComputeModeExclusive, ::cudaComputeModeProhibited or when long running CUDA kernels have filled up the GPU and are blocking new work from starting. They can also be unavailable due to memory constraints on a device that already has active CUDA work being performed.

cudaError_cudaErrorDuplicateSurfaceName
This indicates that multiple surfaces (across separate CUDA source files in the application) share the same string name.

cudaError_cudaErrorDuplicateTextureName
This indicates that multiple textures (across separate CUDA source files in the application) share the same string name.

cudaError_cudaErrorDuplicateVariableName
This indicates that multiple global or constant variables (across separate CUDA source files in the application) share the same string name.

cudaError_cudaErrorECCUncorrectable
This indicates that an uncorrectable ECC error was detected during execution.

cudaError_cudaErrorFileNotFound
This indicates that the file specified was not found.

cudaError_cudaErrorHardwareStackError
Device encountered an error in the call stack during kernel execution, possibly due to stack corruption or exceeding the stack size limit. This leaves the process in an inconsistent state and any further CUDA work will return the same error. To continue using CUDA, the process must be terminated and relaunched.

cudaError_cudaErrorHostMemoryAlreadyRegistered
This error indicates that the memory range passed to ::cudaHostRegister() has already been registered.

cudaError_cudaErrorHostMemoryNotRegistered
This error indicates that the pointer passed to ::cudaHostUnregister() does not correspond to any currently registered memory region.

cudaError_cudaErrorIllegalAddress
The device encountered a load or store instruction on an invalid memory address. This leaves the process in an inconsistent state and any further CUDA work will return the same error. To continue using CUDA, the process must be terminated and relaunched.

cudaError_cudaErrorIllegalInstruction
The device encountered an illegal instruction during kernel execution This leaves the process in an inconsistent state and any further CUDA work will return the same error. To continue using CUDA, the process must be terminated and relaunched.

cudaError_cudaErrorIllegalState
This indicates that a resource required by the API call is not in a valid state to perform the requested operation.

cudaError_cudaErrorIncompatibleDriverContext
This indicates that the current context is not compatible with this the CUDA Runtime. This can only occur if you are using CUDA Runtime/Driver interoperability and have created an existing Driver context using the driver API. The Driver context may be incompatible either because the Driver context was created using an older version of the API, because the Runtime API call expects a primary driver context and the Driver context is not primary, or because the Driver context has been destroyed. Please see \ref CUDART_DRIVER “Interactions with the CUDA Driver API” for more information.

cudaError_cudaErrorInitializationError
The API call failed because the CUDA driver and runtime could not be initialized.

cudaError_cudaErrorInsufficientDriver
This indicates that the installed NVIDIA CUDA driver is older than the CUDA runtime library. This is not a supported configuration. Users should install an updated NVIDIA display driver to allow the application to run.

cudaError_cudaErrorInvalidAddressSpace
While executing a kernel, the device encountered an instruction which can only operate on memory locations in certain address spaces (global, shared, or local), but was supplied a memory address not belonging to an allowed address space. This leaves the process in an inconsistent state and any further CUDA work will return the same error. To continue using CUDA, the process must be terminated and relaunched.

cudaError_cudaErrorInvalidChannelDescriptor
This indicates that the channel descriptor passed to the API call is not valid. This occurs if the format is not one of the formats specified by ::cudaChannelFormatKind, or if one of the dimensions is invalid.

cudaError_cudaErrorInvalidConfiguration
This indicates that a kernel launch is requesting resources that can never be satisfied by the current device. Requesting more shared memory per block than the device supports will trigger this error, as will requesting too many threads or blocks. See ::cudaDeviceProp for more device limitations.

cudaError_cudaErrorInvalidDevice
This indicates that the device ordinal supplied by the user does not correspond to a valid CUDA device.

cudaError_cudaErrorInvalidDeviceFunction
The requested device function does not exist or is not compiled for the proper device architecture.

cudaError_cudaErrorInvalidDevicePointer
This indicates that at least one device pointer passed to the API call is not a valid device pointer. \deprecated This error return is deprecated as of CUDA 10.1.

cudaError_cudaErrorInvalidFilterSetting
This indicates that a non-float texture was being accessed with linear filtering. This is not supported by CUDA.

cudaError_cudaErrorInvalidGraphicsContext
This indicates an error with the OpenGL or DirectX context.

cudaError_cudaErrorInvalidHostPointer
This indicates that at least one host pointer passed to the API call is not a valid host pointer. \deprecated This error return is deprecated as of CUDA 10.1.

cudaError_cudaErrorInvalidKernelImage
This indicates that the device kernel image is invalid.

cudaError_cudaErrorInvalidMemcpyDirection
This indicates that the direction of the memcpy passed to the API call is not one of the types specified by ::cudaMemcpyKind.

cudaError_cudaErrorInvalidNormSetting
This indicates that an attempt was made to read a non-float texture as a normalized float. This is not supported by CUDA.

cudaError_cudaErrorInvalidPc
The device encountered an invalid program counter. This leaves the process in an inconsistent state and any further CUDA work will return the same error. To continue using CUDA, the process must be terminated and relaunched.

cudaError_cudaErrorInvalidPitchValue
This indicates that one or more of the pitch-related parameters passed to the API call is not within the acceptable range for pitch.

cudaError_cudaErrorInvalidPtx
A PTX compilation failed. The runtime may fall back to compiling PTX if an application does not contain a suitable binary for the current device.

cudaError_cudaErrorInvalidResourceHandle
This indicates that a resource handle passed to the API call was not valid. Resource handles are opaque types like ::cudaStream_t and ::cudaEvent_t.

cudaError_cudaErrorInvalidSource
This indicates that the device kernel source is invalid.

cudaError_cudaErrorInvalidSurface
This indicates that the surface passed to the API call is not a valid surface.

cudaError_cudaErrorInvalidSymbol
This indicates that the symbol name/identifier passed to the API call is not a valid name or identifier.

cudaError_cudaErrorInvalidTexture
This indicates that the texture passed to the API call is not a valid texture.

cudaError_cudaErrorInvalidTextureBinding
This indicates that the texture binding is not valid. This occurs if you call ::cudaGetTextureAlignmentOffset() with an unbound texture.

cudaError_cudaErrorInvalidValue
This indicates that one or more of the parameters passed to the API call is not within an acceptable range of values.

cudaError_cudaErrorJitCompilerNotFound
This indicates that the PTX JIT compiler library was not found. The JIT Compiler library is used for PTX compilation. The runtime may fall back to compiling PTX if an application does not contain a suitable binary for the current device.

cudaError_cudaErrorLaunchFailure
An exception occurred on the device while executing a kernel. Common causes include dereferencing an invalid device pointer and accessing out of bounds shared memory. Less common cases can be system specific - more information about these cases can be found in the system specific user guide. This leaves the process in an inconsistent state and any further CUDA work will return the same error. To continue using CUDA, the process must be terminated and relaunched.

cudaError_cudaErrorLaunchFileScopedSurf
This error indicates that a grid launch did not occur because the kernel uses file-scoped surfaces which are unsupported by the device runtime. Kernels launched via the device runtime only support surfaces created with the Surface Object API’s.

cudaError_cudaErrorLaunchFileScopedTex
This error indicates that a grid launch did not occur because the kernel uses file-scoped textures which are unsupported by the device runtime. Kernels launched via the device runtime only support textures created with the Texture Object API’s.

cudaError_cudaErrorLaunchIncompatibleTexturing
This error indicates a kernel launch that uses an incompatible texturing mode.

cudaError_cudaErrorLaunchMaxDepthExceeded
This error indicates that a device runtime grid launch did not occur because the depth of the child grid would exceed the maximum supported number of nested grid launches.

cudaError_cudaErrorLaunchOutOfResources
This indicates that a launch did not occur because it did not have appropriate resources. Although this error is similar to ::cudaErrorInvalidConfiguration, this error usually indicates that the user has attempted to pass too many arguments to the device kernel, or the kernel launch specifies too many threads for the kernel’s register count.

cudaError_cudaErrorLaunchPendingCountExceeded
This error indicates that a device runtime grid launch failed because the launch would exceed the limit ::cudaLimitDevRuntimePendingLaunchCount. For this launch to proceed successfully, ::cudaDeviceSetLimit must be called to set the ::cudaLimitDevRuntimePendingLaunchCount to be higher than the upper bound of outstanding launches that can be issued to the device runtime. Keep in mind that raising the limit of pending device runtime launches will require the runtime to reserve device memory that cannot be used for user allocations.

cudaError_cudaErrorLaunchTimeout
This indicates that the device kernel took too long to execute. This can only occur if timeouts are enabled - see the device property \ref ::cudaDeviceProp::kernelExecTimeoutEnabled “kernelExecTimeoutEnabled” for more information. This leaves the process in an inconsistent state and any further CUDA work will return the same error. To continue using CUDA, the process must be terminated and relaunched.

cudaError_cudaErrorMapBufferObjectFailed
This indicates that the buffer object could not be mapped.

cudaError_cudaErrorMemoryAllocation
The API call failed because it was unable to allocate enough memory to perform the requested operation.

cudaError_cudaErrorMemoryValueTooLarge
This indicated that an emulated device pointer exceeded the 32-bit address range. \deprecated This error return is deprecated as of CUDA 3.1. Device emulation mode was removed with the CUDA 3.1 release.

cudaError_cudaErrorMisalignedAddress
The device encountered a load or store instruction on a memory address which is not aligned. This leaves the process in an inconsistent state and any further CUDA work will return the same error. To continue using CUDA, the process must be terminated and relaunched.

cudaError_cudaErrorMissingConfiguration
The device function being invoked (usually via ::cudaLaunchKernel()) was not previously configured via the ::cudaConfigureCall() function.

cudaError_cudaErrorMixedDeviceExecution
Mixing of device and device emulation code was not allowed. \deprecated This error return is deprecated as of CUDA 3.1. Device emulation mode was removed with the CUDA 3.1 release.

cudaError_cudaErrorNoDevice
This indicates that no CUDA-capable devices were detected by the installed CUDA driver.

cudaError_cudaErrorNoKernelImageForDevice
This indicates that there is no kernel image available that is suitable for the device. This can occur when a user specifies code generation options for a particular CUDA source file that do not include the corresponding device configuration.

cudaError_cudaErrorNotMapped
This indicates that a resource is not mapped.

cudaError_cudaErrorNotMappedAsArray
This indicates that a mapped resource is not available for access as an array.

cudaError_cudaErrorNotMappedAsPointer
This indicates that a mapped resource is not available for access as a pointer.

cudaError_cudaErrorNotPermitted
This error indicates the attempted operation is not permitted.

cudaError_cudaErrorNotReady
This indicates that asynchronous operations issued previously have not completed yet. This result is not actually an error, but must be indicated differently than ::cudaSuccess (which indicates completion). Calls that may return this value include ::cudaEventQuery() and ::cudaStreamQuery().

cudaError_cudaErrorNotSupported
This error indicates the attempted operation is not supported on the current system or device.

cudaError_cudaErrorNotYetImplemented
This indicates that the API call is not yet implemented. Production releases of CUDA will never return this error. \deprecated This error return is deprecated as of CUDA 4.1.

cudaError_cudaErrorNvlinkUncorrectable
This indicates that an uncorrectable NVLink error was detected during the execution.

cudaError_cudaErrorOperatingSystem
This error indicates that an OS call failed.

cudaError_cudaErrorPeerAccessAlreadyEnabled
This error indicates that a call to ::cudaDeviceEnablePeerAccess() is trying to re-enable peer addressing on from a context which has already had peer addressing enabled.

cudaError_cudaErrorPeerAccessNotEnabled
This error indicates that ::cudaDeviceDisablePeerAccess() is trying to disable peer addressing which has not been enabled yet via ::cudaDeviceEnablePeerAccess().

cudaError_cudaErrorPeerAccessUnsupported
This error indicates that P2P access is not supported across the given devices.

cudaError_cudaErrorPriorLaunchFailure
This indicated that a previous kernel launch failed. This was previously used for device emulation of kernel launches. \deprecated This error return is deprecated as of CUDA 3.1. Device emulation mode was removed with the CUDA 3.1 release.

cudaError_cudaErrorProfilerAlreadyStarted
\deprecated This error return is deprecated as of CUDA 5.0. It is no longer an error to call cudaProfilerStart() when profiling is already enabled.

cudaError_cudaErrorProfilerAlreadyStopped
\deprecated This error return is deprecated as of CUDA 5.0. It is no longer an error to call cudaProfilerStop() when profiling is already disabled.

cudaError_cudaErrorProfilerDisabled
This indicates profiler is not initialized for this run. This can happen when the application is running with external profiling tools like visual profiler.

cudaError_cudaErrorProfilerNotInitialized
\deprecated This error return is deprecated as of CUDA 5.0. It is no longer an error to attempt to enable/disable the profiling via ::cudaProfilerStart or ::cudaProfilerStop without initialization.

cudaError_cudaErrorSetOnActiveProcess
This indicates that the user has called ::cudaSetValidDevices(), ::cudaSetDeviceFlags(), ::cudaD3D9SetDirect3DDevice(), ::cudaD3D10SetDirect3DDevice, ::cudaD3D11SetDirect3DDevice(), or ::cudaVDPAUSetVDPAUDevice() after initializing the CUDA runtime by calling non-device management operations (allocating memory and launching kernels are examples of non-device management operations). This error can also be returned if using runtime/driver interoperability and there is an existing ::CUcontext active on the host thread.

cudaError_cudaErrorSharedObjectInitFailed
This indicates that initialization of a shared object failed.

cudaError_cudaErrorSharedObjectSymbolNotFound
This indicates that a link to a shared object failed to resolve.

cudaError_cudaErrorStartupFailure
This indicates an internal startup failure in the CUDA runtime.

cudaError_cudaErrorStreamCaptureImplicit
The operation would have resulted in a disallowed implicit dependency on a current capture sequence from cudaStreamLegacy.

cudaError_cudaErrorStreamCaptureInvalidated
The current capture sequence on the stream has been invalidated due to a previous error.

cudaError_cudaErrorStreamCaptureIsolation
A dependency would have been created which crosses the capture sequence boundary. Only implicit in-stream ordering dependencies are allowed to cross the boundary.

cudaError_cudaErrorStreamCaptureMerge
The operation would have resulted in a merge of two independent capture sequences.

cudaError_cudaErrorStreamCaptureUnjoined
The capture sequence contains a fork that was not joined to the primary stream.

cudaError_cudaErrorStreamCaptureUnmatched
The capture was not initiated in this stream.

cudaError_cudaErrorStreamCaptureUnsupported
The operation is not permitted when the stream is capturing.

cudaError_cudaErrorStreamCaptureWrongThread
A stream capture sequence not initiated with the ::cudaStreamCaptureModeRelaxed argument to ::cudaStreamBeginCapture was passed to ::cudaStreamEndCapture in a different thread.

cudaError_cudaErrorSymbolNotFound
This indicates that a named symbol was not found. Examples of symbols are global/constant variable names, texture names, and surface names.

cudaError_cudaErrorSyncDepthExceeded
This error indicates that a call to ::cudaDeviceSynchronize made from the device runtime failed because the call was made at grid depth greater than than either the default (2 levels of grids) or user specified device limit ::cudaLimitDevRuntimeSyncDepth. To be able to synchronize on launched grids at a greater depth successfully, the maximum nested depth at which ::cudaDeviceSynchronize will be called must be specified with the ::cudaLimitDevRuntimeSyncDepth limit to the ::cudaDeviceSetLimit api before the host-side launch of a kernel using the device runtime. Keep in mind that additional levels of sync depth require the runtime to reserve large amounts of device memory that cannot be used for user allocations.

cudaError_cudaErrorSynchronizationError
This indicated that a synchronization operation had failed. This was previously used for some device emulation functions. \deprecated This error return is deprecated as of CUDA 3.1. Device emulation mode was removed with the CUDA 3.1 release.

cudaError_cudaErrorSystemDriverMismatch
This error indicates that there is a mismatch between the versions of the display driver and the CUDA driver. Refer to the compatibility documentation for supported versions.

cudaError_cudaErrorSystemNotReady
This error indicates that the system is not yet ready to start any CUDA work. To continue using CUDA, verify the system configuration is in a valid state and all required driver daemons are actively running. More information about this error can be found in the system specific user guide.

cudaError_cudaErrorTextureFetchFailed
This indicated that a texture fetch was not able to be performed. This was previously used for device emulation of texture operations. \deprecated This error return is deprecated as of CUDA 3.1. Device emulation mode was removed with the CUDA 3.1 release.

cudaError_cudaErrorTextureNotBound
This indicated that a texture was not bound for access. This was previously used for device emulation of texture operations. \deprecated This error return is deprecated as of CUDA 3.1. Device emulation mode was removed with the CUDA 3.1 release.

cudaError_cudaErrorTooManyPeers
This error indicates that the hardware resources required to enable peer access have been exhausted for one or more of the devices passed to ::cudaEnablePeerAccess().

cudaError_cudaErrorUnknown
This indicates that an unknown internal error has occurred.

cudaError_cudaErrorUnmapBufferObjectFailed
This indicates that the buffer object could not be unmapped.

cudaError_cudaErrorUnsupportedLimit
This indicates that the ::cudaLimit passed to the API call is not supported by the active device.

cudaError_cudaSuccess
The API call returned with no errors. In the case of query calls, this also means that the operation being queried is complete (see ::cudaEventQuery() and ::cudaStreamQuery()).

cusparseStatus_t_CUSPARSE_STATUS_ALLOC_FAILED

cusparseStatus_t_CUSPARSE_STATUS_ARCH_MISMATCH

cusparseStatus_t_CUSPARSE_STATUS_EXECUTION_FAILED

cusparseStatus_t_CUSPARSE_STATUS_INTERNAL_ERROR

cusparseStatus_t_CUSPARSE_STATUS_INVALID_VALUE

cusparseStatus_t_CUSPARSE_STATUS_MAPPING_ERROR

cusparseStatus_t_CUSPARSE_STATUS_MATRIX_TYPE_NOT_SUPPORTED

cusparseStatus_t_CUSPARSE_STATUS_NOT_INITIALIZED

cusparseStatus_t_CUSPARSE_STATUS_SUCCESS

cusparseStatus_t_CUSPARSE_STATUS_ZERO_PIVOT

Functions

BTBT_MVM⚠

BTBT_MVM_i⚠

BTBT_MVM_ii⚠

BTBT_MVM_iii⚠

BTBT_MVM_iv⚠

BTBT_MVM_v⚠

BTBT_cleanup⚠

BTBT_info⚠

BTBT_setup⚠

BTBT_setup1⚠

BTBT_setup2⚠

GBTBT_MVM⚠

GBTBT_cleanup⚠

GBTBT_info⚠

GBTBT_setup⚠

GBTBT_setup1⚠

GBTBT_setup2⚠

GBTBT_setup3⚠

LMMSE_cleanup⚠

LMMSE_estimation⚠

LMMSE_estimation1⚠

LMMSE_estimation_old⚠

LMMSE_estimation_old1⚠

LMMSE_reset⚠

LMMSE_set_phase_est_ptr⚠

LMMSE_setup⚠

LMMSE_setup1⚠

LMMSE_setup2⚠

LMMSE_setup3⚠

LMMSE_setup4⚠

LMMSE_setup5⚠

LMMSE_toFile⚠

aaStats_cleanup⚠

aaStats_info⚠

aaStats_setup⚠

aaStats_toFile⚠

aaStats_variance⚠

aperture_cleanup⚠

aperture_setup⚠

aperture_setup1⚠

aperture_setup2⚠

aperture_setup_GMT_M1⚠

aperture_vignetting⚠

atmosphere_cleanup⚠

atmosphere_get_phase_screen⚠

atmosphere_get_phase_screen1⚠

atmosphere_get_phase_screen2⚠

atmosphere_get_phase_screen3⚠

atmosphere_get_phase_screen4⚠

atmosphere_get_phase_screen5⚠

atmosphere_get_phase_screen6⚠

atmosphere_get_phase_screen7⚠

atmosphere_get_phase_screen_circ_centroids⚠

atmosphere_get_phase_screen_circ_uplink_centroids⚠

atmosphere_get_phase_screen_gradient⚠

atmosphere_get_phase_screen_gradient1⚠

atmosphere_get_phase_screen_gradient2⚠

atmosphere_get_phase_screen_gradient3⚠

atmosphere_get_phase_screen_gradient4⚠

atmosphere_get_phase_screen_gradient5⚠

atmosphere_get_phase_screen_gradient_rolling_shutter⚠

atmosphere_gmt_set_eph⚠

atmosphere_gmt_set_id⚠

atmosphere_gmt_setup⚠

atmosphere_gmt_setup1⚠

atmosphere_gmt_setup2⚠

atmosphere_gmt_setup3⚠

atmosphere_gmt_setup4⚠

atmosphere_gmt_setup5⚠

atmosphere_gmt_setup6⚠

atmosphere_info⚠

atmosphere_rayTracing⚠

atmosphere_rayTracing1⚠

atmosphere_reset⚠

atmosphere_save_layer_phasescreens⚠

atmosphere_setup⚠

atmosphere_setup1⚠

atmosphere_setup2⚠

bundle_cleanup⚠

bundle_get_chief_coordinates⚠

bundle_get_chief_directions⚠

bundle_get_chief_optical_path_length⚠

bundle_get_coordinates⚠

bundle_get_directions⚠

bundle_get_n_iteration⚠

bundle_get_optical_path_difference⚠

bundle_get_optical_path_difference1⚠

bundle_get_optical_path_length⚠

bundle_get_sphere_origins⚠

bundle_get_vignetting⚠

bundle_gmt_m2_baffle⚠

bundle_gmt_truss_onaxis⚠

bundle_setup⚠

bundle_setup1⚠

bundle_setup2⚠

bundle_setup_free⚠

bundle_setup_free1⚠

bundle_to_focal_plane⚠

bundle_to_sphere⚠

bundle_to_sphere1⚠

bundle_to_z_plane⚠

centroiding_cleanup⚠

centroiding_filter⚠

centroiding_fried_geometry⚠

centroiding_get_data⚠

centroiding_get_data1⚠

centroiding_get_data2⚠

centroiding_get_data3⚠

centroiding_noisify⚠

centroiding_reset⚠

centroiding_setup⚠

centroiding_show_centroids⚠

centroiding_show_flux⚠

complex_amplitude_add_phase⚠

complex_amplitude_add_same_phase⚠

complex_amplitude_cleanup⚠

complex_amplitude_finite_difference⚠

complex_amplitude_finite_difference1⚠

complex_amplitude_gradient_average⚠

complex_amplitude_gradient_average1⚠

complex_amplitude_masked⚠

complex_amplitude_masked1⚠

complex_amplitude_reset⚠

complex_amplitude_reset1⚠

complex_amplitude_reset_amplitude⚠

complex_amplitude_reset_phase⚠

complex_amplitude_reset_phase1⚠

complex_amplitude_rms⚠

complex_amplitude_segments_gradient_average⚠

complex_amplitude_segments_gradient_averageFast⚠

complex_amplitude_setup⚠

complex_amplitude_setup1⚠

complex_amplitude_show_amplitude⚠

complex_amplitude_show_amplitude1⚠

complex_amplitude_show_phase⚠

complex_amplitude_show_phase1⚠

conic_cleanup⚠

conic_info⚠

conic_setup⚠

conic_setup1⚠

conic_setup2⚠

conic_setup3⚠

conic_setup4⚠

conic_trace⚠

coordinate_system_cleanup⚠

coordinate_system_info⚠

coordinate_system_info_details⚠

coordinate_system_setup⚠

coordinate_system_setup1⚠

coordinate_system_setup2⚠

coordinate_system_setup3⚠

coordinate_system_update⚠

coordinate_system_update1⚠

dev2host⚠

dev2host_int⚠

geometricShackHartmann_analyze⚠

geometricShackHartmann_calibrate⚠

geometricShackHartmann_cleanup⚠

geometricShackHartmann_folded_slopes⚠

geometricShackHartmann_get_valid_reference_slopes⚠

geometricShackHartmann_get_valid_slopes⚠

geometricShackHartmann_get_valid_slopes_norm⚠

geometricShackHartmann_identify_valid_lenslet⚠

geometricShackHartmann_masked_slopes⚠

geometricShackHartmann_process⚠

geometricShackHartmann_propagate⚠

geometricShackHartmann_propagate1⚠

geometricShackHartmann_reset⚠

geometricShackHartmann_set_reference_slopes⚠

geometricShackHartmann_setup⚠

geqrf⚠

gmt_m1_blocking⚠

gmt_m1_cleanup⚠

gmt_m1_edge_sensors_data⚠

gmt_m1_global_tiptilt⚠

gmt_m1_keep⚠

gmt_m1_locate⚠

gmt_m1_preset⚠

gmt_m1_remove⚠

gmt_m1_reset⚠

gmt_m1_set_reflectivity⚠

gmt_m1_setup⚠

gmt_m1_setup1⚠

gmt_m1_setup2⚠

gmt_m1_test_ray_tracing⚠

gmt_m1_trace⚠

gmt_m1_traceall⚠

gmt_m1_track⚠

gmt_m1_update⚠

gmt_m1_update_conic_c⚠

gmt_m1_update_conic_k⚠

gmt_m2_blocking⚠

gmt_m2_cleanup⚠

gmt_m2_coma_neutral⚠

gmt_m2_global_tiptilt⚠

gmt_m2_keep⚠

gmt_m2_locate⚠

gmt_m2_pointing_neutral⚠

gmt_m2_remove⚠

gmt_m2_reset⚠

gmt_m2_set_reflectivity⚠

gmt_m2_setup⚠

gmt_m2_setup1⚠

gmt_m2_setup2⚠

gmt_m2_test_ray_tracing⚠

gmt_m2_trace⚠

gmt_m2_traceall⚠

gmt_m2_track⚠

gmt_m2_update⚠

gmt_m2_update_conic_c⚠

gmt_m2_update_conic_k⚠

gpu_double_dev2host⚠

gpu_double_dev_malloc⚠

gpu_double_free_dev⚠

gpu_double_host2dev⚠

gpu_double_mv⚠

gpu_double_qr⚠

gpu_double_reset⚠

gpu_double_setup⚠

gpu_double_setup1⚠

gpu_float_axpy⚠

gpu_float_dev2dev⚠

gpu_float_dev2host⚠

gpu_float_dev_malloc⚠

gpu_float_free_dev⚠

gpu_float_host2dev⚠

gpu_float_mv⚠

gpu_float_qr⚠

gpu_float_qr_solve⚠

gpu_float_reset⚠

gpu_float_scale⚠

gpu_float_setup⚠

gpu_float_setup1⚠

host2dev⚠

host2dev_char⚠

imaging_cleanup⚠

imaging_cleanupSegmentPistonSensor⚠

imaging_frame2file⚠

imaging_info⚠

imaging_noiseless_readout⚠

imaging_propagate⚠

imaging_propagateNoOverlap⚠

imaging_propagateNoOverlapBare⚠

imaging_propagateNoOverlapSPS⚠

imaging_propagateTT7⚠

imaging_propagateTT71⚠

imaging_propagateThroughFieldStop⚠

imaging_propagateThroughModulatedPyramid⚠

imaging_propagateThroughPyramid⚠

imaging_propagate_cpx⚠

imaging_readout⚠

imaging_readout1⚠

imaging_reset⚠

imaging_reset_rng⚠

imaging_set_pointing_direction⚠

imaging_setup⚠

imaging_setup1⚠

imaging_setup2⚠

imaging_setup3⚠

imaging_setupSegmentPistonSensor⚠

imaging_show_frame⚠

imaging_show_frame1⚠

imaging_strehl_ratio⚠

intersect⚠

iterativeSolvers_cg⚠

iterativeSolvers_cg_setup⚠

iterativeSolvers_cleanup⚠

iterativeSolvers_lanczos_step⚠

iterativeSolvers_minres_choi⚠

iterativeSolvers_minres_setup⚠

iterativeSolvers_minres_vorst⚠

iterativeSolvers_minres_vorst1⚠

iterativeSolvers_minres_vorst2⚠

iterativeSolvers_minres_vorst3⚠

iterativeSolvers_pcg⚠

iterativeSolvers_pcg_setup⚠

iterativeSolvers_pminres_vorst⚠

iterativeSolvers_sym_ortho⚠

layer_setup⚠

mask_add⚠

mask_add1⚠

mask_add2⚠

mask_add3⚠

mask_alter⚠

mask_cleanup⚠

mask_reset⚠

mask_set_filter⚠

mask_set_filter_quiet⚠

mask_set_gmt_piston⚠

mask_set_index⚠

mask_setup⚠

mask_setup1⚠

mask_setup2⚠

mask_setup3⚠

mask_setup_GMT⚠

mask_setup_circular⚠

mask_setup_circular1⚠

mask_setup_circular2⚠

modes_cleanup⚠

modes_load⚠

modes_load1⚠

modes_reset_modes⚠

modes_setup⚠

modes_setup1⚠

modes_update⚠

ormqr⚠

paStats_MVM⚠

paStats_cleanup⚠

paStats_info⚠

paStats_setup⚠

paStats_setup1⚠

paStats_toFile⚠

paStats_variance⚠

profile_cleanup⚠

profile_setup⚠

pssn_B⚠

pssn_C⚠

pssn_O⚠

pssn_O0⚠

pssn___otf__⚠

pssn_atm_otf⚠

pssn_cleanup⚠

pssn_eval⚠

pssn_eval1⚠

pssn_oeval⚠

pssn_oeval1⚠

pssn_otf⚠

pssn_setup⚠

pssn_xotf⚠

pyramid_cleanup⚠

pyramid_setup⚠

reflect⚠

refract⚠

set_device⚠

shackHartmann_analyze⚠

shackHartmann_calibrate⚠

shackHartmann_cleanup⚠

shackHartmann_get_frame_dev_ptr⚠

shackHartmann_get_valid_reference_slopes⚠

shackHartmann_get_valid_slopes⚠

shackHartmann_get_valid_slopes_norm⚠

shackHartmann_identify_valid_lenslet⚠

shackHartmann_masked_slopes⚠

shackHartmann_process⚠

shackHartmann_propagate⚠

shackHartmann_propagate1⚠

shackHartmann_set_reference_slopes⚠

shackHartmann_setup⚠

shackHartmann_update_lenslet⚠

source_cleanup⚠

source_copy_magnitude⚠

source_info⚠

source_n_background_photon⚠

source_n_photon⚠

source_n_photon1⚠

source_opd2phase⚠

source_opd2phase1⚠

source_optical_transfer_function⚠

source_phase2file⚠

source_reset_rays⚠

source_reset_rays1⚠

source_setup⚠

source_setup1⚠

source_setup2⚠

source_setup3⚠

source_setup4⚠

source_setup5⚠

source_setup6⚠

source_setup7⚠

source_setup8⚠

source_setup9⚠

source_setup_chief⚠

source_spectral_bandwidth⚠

source_update_directions⚠

source_update_magnitude⚠

source_wavelength⚠

source_wavelength_micron⚠

source_wavenumber⚠

stats_cleanup⚠

stats_diff_std⚠

stats_diff_std1⚠

stats_diff_var⚠

stats_diff_var1⚠

stats_mean⚠

stats_mean1⚠

stats_setup⚠

stats_std⚠

stats_std1⚠

stats_var⚠

stats_var1⚠

transform_to_R⚠

transform_to_R1⚠

transform_to_S⚠

transform_to_S1⚠

vector_angle⚠

vector_dot⚠

vector_left_cross⚠

vector_mag⚠

vector_mag1⚠

vector_norm⚠

vector_rho2⚠

vector_rho2_shift⚠

vector_right_cross⚠

vector_unit⚠

zernikeS_cleanup⚠

zernikeS_projection⚠

zernikeS_setup⚠

zernikeS_setup1⚠

zernikeS_setup2⚠

zernikeS_setup3⚠

zernikeS_surface⚠

zernikeS_surface1⚠

zernikeS_surface_and_derivatives⚠

zernikeS_surface_and_derivatives1⚠

zernikeS_surface_derivative_coefs⚠

zernikeS_surface_derivatives⚠

zernikeS_update⚠

Type Definitions

cublasHandle_t

cublasStatus_t

cudaError
CUDA error types

cudaEvent_t
CUDA event types

cufftHandle

curandState

cusolverDnHandle_t

cusparseHandle_t

cusparseMatDescr_t

cusparseStatus_t

rtd