Macro crseo::ceo

macro_rules! ceo {
    ($element:ident) => { ... };
    ($element:ident, $($arg:ident = [$($val:expr),*]),*) => { ... };
    ($element:ident:$model:ident, $($arg:ident = [$($val:expr),*]),*) => { ... };
    ($element:ident:$model:ident, $($arg:ident = [$($val:expr),*]),*) => { ... };
}

CEO macro builder

One macro to rule them all, one macro to find them, one macro to bring them all and in the darkness bind them all

Examples

GMT

use ceo::ceo;
let gmt = ceo!(GMT, set_m1_n_mode = [27], set_m2_n_mode = [123]);

Geometric Shack-Hartmann

use ceo::ceo;
let mut wfs = ceo!(
    SHACKHARTMANN:Geometric,
    set_n_sensor = [1],
    set_lenslet_array = [48, 16, 25.5 / 48f64]
);
let mut src = ceo!(SOURCE, set_pupil_sampling = [48 * 16 + 1]);
let mut gmt = ceo!(GMT);
src.through(&mut gmt).xpupil().through(&mut wfs);
println!("WFE RMS: {:.3}nm", src.wfe_rms_10e(-9)[0]);

Diffractive Shack-Hartmann

/*use ceo::ceo;
let mut wfs = ceo!(
    SHACKHARTMANN<Diffractive>,
    set_n_sensor = [1],
    set_lenslet_array = [48, 16, 25.5 / 48f64],
    set_detector = [8, Some(24), None]
);
let mut src = ceo!(SOURCE, set_pupil_sampling = [48 * 16 + 1]);
let mut gmt = ceo!(GMT);
src.through(&mut gmt).xpupil().through(&mut wfs);
println!("WFE RMS: {:.3}nm", src.wfe_rms_10e(-9)[0]);*/