Struct crseo::shackhartmann::ShackHartmann

pub struct ShackHartmann<S: Model> {
    pub _c_: S,
    pub n_side_lenslet: i32,
    pub n_px_lenslet: i32,
    pub d: f64,
    pub n_sensor: i32,
    pub n_centroids: i32,
    pub centroids: Cu<Single>,
}

shackhartmann wrapper

Fields
_c_: Sn_side_lenslet: i32

The size of the square lenslet array

n_px_lenslet: i32

The number of pixel per lenslet in the telescope pupil

d: f64

The lenslet array pitch [m]

n_sensor: i32

The number of WFS

n_centroids: i32

The total number of centroids

centroids: Cu<Single>

The centroids

Implementations

impl<S: Model> ShackHartmann<S>

pub fn new(
n_sensor: i32,
n_side_lenslet: i32,
n_px_lenslet: i32,
d: f64
) -> ShackHartmann<S>

Creates a new ShackHartmann as either Geometric or Diffractive type

n_sensor - the number of WFS

n_side_lenslet - the size of the square lenslet array

n_px_lenslet - the number of pixel per lenslet in the telescope pupil

d - the lenslet pitch [m]

pub fn guide_stars(&self) -> SOURCE

pub fn drop_sh(&mut self)

impl ShackHartmann<Geometric>

pub fn build(&mut self) -> &mut Self

Initializes the ShackHartmann WFS

pub fn guide_star_args(&self) -> (i32, f64, i32)

pub fn new_guide_stars(&self) -> Source

pub fn process(&mut self) -> &mut Self

pub fn get_data(&mut self) -> Cu<Single>

pub fn filter(&mut self, lenslet_mask: &mut Mask) -> Cu<Single>

pub fn fold_into(&mut self, data: &mut Cu<Single>, lenslet_mask: &mut Mask)

pub fn n_valid_lenslet(&mut self) -> usize

pub fn reset(&mut self) -> &mut Self

pub fn lenset_mask(&mut self) -> Cu<Single>

pub fn lenlet_flux(&mut self) -> Cu<Single>

pub fn as_raw_mut_ptr(&mut self) -> &mut Geometric

impl ShackHartmann<Diffractive>

pub fn build(
&mut self,
n_px_framelet: i32,
n_px_imagelet: Option<i32>,
osf: Option<i32>
) -> &mut Self

pub fn new_guide_stars(&self) -> Source

pub fn process(&mut self) -> &mut Self

pub fn readout(
&mut self,
exposure_time: f32,
readout_noise_rms: f32,
n_background_photon: f32,
noise_factor: f32
) -> &mut Self

pub fn reset(&mut self) -> &mut Self