Module crseo::source

CEO source wrapper

Provides a structure Source that is a wrapper for CEO source C++ structure. Source is instantiated and initialized with the SOURCE builder

Examples

on-axis source with default parameters

use ceo::ceo;
// Creates a source with default parameters
let mut src = ceo!(SOURCE);

3 sources evenly spread on a ring with a 8 arcminute radius

use ceo::{ceo, Conversion};
let mut src = ceo!(SOURCE, size = [3] , on_ring = [8f32.from_arcmin()]);

Structs

SOURCE
Source builder

Source
source wrapper

Traits

Propagation
A system that mutates Source arguments should implement the Propagation trait