Struct crseo::source::Source

#[repr(C)]pub struct Source {
    pub size: i32,
    pub pupil_size: f64,
    pub pupil_sampling: i32,
    pub _wfe_rms: Vec<f32>,
    pub _phase: Vec<f32>,
    pub zenith: Vec<f32>,
    pub azimuth: Vec<f32>,
    pub magnitude: Vec<f32>,
    // some fields omitted
}

source wrapper

Fields
size: i32

The number of sources

pupil_size: f64

The diameter of the entrance pupil [m]

pupil_sampling: i32

The sampling of the entrance pupil [px]

_wfe_rms: Vec<f32>_phase: Vec<f32>zenith: Vec<f32>azimuth: Vec<f32>magnitude: Vec<f32>
Implementations

impl Source

pub fn empty() -> Source

Creates and empty Source

pub fn new(size: i32, pupil_size: f64, pupil_sampling: i32) -> Source

Creates a new Source with the arguments:

pupil_size - the diameter of the entrance pupil [m]

pupil_sampling - the sampling of the entrance pupil [px]

pub fn from(args: (i32, f64, i32)) -> Source

pub fn build(
&mut self,
band: &str,
zenith: Vec<f32>,
azimuth: Vec<f32>,
magnitude: Vec<f32>
) -> &mut Self

Sets the Source parameters:

band - the photometric band: Vs, V, R, I, J, H, K or R+I

zenith - the zenith angle [rd]

azimuth - the azimuth angle [rd]

magnitude - the magnitude at the specified photometric band

pub fn as_raw_mut_ptr(&mut self) -> &mut source

pub fn get_photometric_band(&self) -> String

Returns the Source photometric band

pub fn wavelength(&mut self) -> f64

Returns the Source wavelength [m]

pub fn fwhm(&mut self, value: f64)

Sets the Source full width at half maximum in un-binned detector pixel

pub fn rotate_rays(&mut self, angle: f64)

Set the pupil rotation angle [degree]

pub fn xpupil(&mut self) -> &mut Self

Copies the optical path difference from ray tracing into the wavefront phase argument, this usually takes place after ray tracing to the exit pupil

pub fn opd2phase(&mut self) -> &mut Self

pub fn wfe_rms(&mut self) -> Vec<f32>

Returns the wavefront error root mean square [m]

pub fn wfe_rms_f64(&mut self) -> Vec<f64>

pub fn wfe_rms_10e(&mut self, exp: i32) -> Vec<f32>

Returns the wavefront error root mean square [m]x10^-exp

pub fn gradients(&mut self) -> Vec<f32>

pub fn segment_wfe_rms(&mut self) -> Vec<f32>

pub fn segment_wfe_rms_10e(&mut self, exp: i32) -> Vec<f32>

pub fn segment_piston(&mut self) -> Vec<f32>

pub fn segment_piston_10e(&mut self, exp: i32) -> Vec<f32>

pub fn segment_mask(&mut self) -> Vec<i32>

pub fn segments_gradients(&mut self) -> Vec<Vec<f32>>

Returns the x and y gradient of the wavefront in average over each of the GMT segments

pub fn lenslet_gradients(
&mut self,
n_lenslet: i32,
_lenslet_size: f64,
data: &mut Centroiding
)

Returns the x and y gradient of the wavefront in average over each lenslet of a n_lensletxn_lenslet array, the gradients are saved in Centroiding

pub fn reset(&mut self)

Resets the rays and the wavefront to their original state

pub fn update(&mut self, zenith: Vec<f64>, azimuth: Vec<f64>)

Updates the zenith and azimuth of the Source

pub fn add(&mut self, phase: &mut Cu<Single>) -> &mut Self

Adds phase to the Source wavefront

pub fn sub(&mut self, phase: &mut Cu<Single>) -> &mut Self

pub fn add_same(&mut self, phase: &mut Cu<Single>) -> &mut Self

Adds phase to the Source wavefront

pub fn phase(&mut self) -> &Vec<f32>

Returns the wavefront phase [m] in the exit pupil of the telescope

pub fn phase_as_ptr(&mut self) -> Cu<Single>

pub fn amplitude(&mut self) -> Vec<f32>

Returns the wavefront amplitude in the exit pupil of the telescope

pub fn fluxlet(&mut self, n_let: usize) -> Vec<f32>

Returns the flux integrated in n_letXn_let bins

pub fn masklet(&mut self, n_let: usize, flux_threshold: f32) -> Vec<i8>

Returns a binary mask where the flux integrated in n_letXn_let bins is greater or equal to the maximum integrated flux X flux_threshold

pub fn through<T: Propagation>(&mut self, system: &mut T) -> &mut Self

Propagates a Source through a system that implements the Propagation trait

pub fn n_photon(&mut self) -> Vec<f32>

Returns the number of photon [m^-2.s^-1]

pub fn light_collecting_area(&self) -> f32

Returns the light collecting area