Struct crseo::Centroiding

pub struct Centroiding {
    pub n_lenslet_total: u32,
    pub n_centroids: u32,
    pub units: f32,
    pub valid_lenslets: Vec<i8>,
    pub n_valid_lenslet: u32,
    pub centroids: Vec<f32>,
    // some fields omitted
}

Wrapper for CEO centroiding

Fields
n_lenslet_total: u32

The total number of lenslets

n_centroids: u32

The number of centroids

units: f32

The centroid units, default: 1 (pixel)

valid_lenslets: Vec<i8>

The valid lenslet mask

n_valid_lenslet: u32

The number of valid lenslet

centroids: Vec<f32>

The centroids

Implementations

impl Centroiding

pub fn new() -> Centroiding

Creates a new Centroiding

pub fn build(&mut self, n_lenslet: u32, data_units: Option<f64>) -> &mut Self

Sets the Centroiding parameters:

n_lenslet - the size of the square lenslet array

data_units - the centroids units

pub fn process(
&mut self,
sensor: &Imaging,
reference: Option<&Centroiding>
) -> &mut Self

Computes the centroids from the sensor image; optionally, a Centroiding reference may be provided that offsets the centroids and sets the valid_lenslets

pub fn grab(&mut self) -> &mut Self

grabs the centroids from the GPU

pub fn valids(&self, some_valid_lenslets: Option<&Vec<i8>>) -> Vec<f32>

returns the valid centroids i.e. the centroids that corresponds to a non-zero entry in the valid_lenslets mask; if some_valid_lenslet is given, then it supersedes any preset valid_lenset

pub fn lenslet_flux(&mut self) -> &Vec<f32>

returns the flux of each lenslet

pub fn integrated_flux(&mut self) -> f32

returns the sum of the flux of all the lenslets

pub fn valid_lenslets(
&mut self,
some_flux_threshold: Option<f64>,
some_valid_lenslets: Option<Vec<i8>>
) -> u32

Computes the valid lenslets and return the number of valid lenslets; the valid lenslets are computed based on the maximum flux threshold or a given valid lenslets mask

pub fn __ceo__(&mut self) -> (&centroiding, &mask)

pub fn __mut_ceo__(&mut self) -> (&mut centroiding, &mut mask)